requirements.txt
# Web framework
flask>=3.0.0
werkzeug>=3.0.0

# Document parsing (existing pipeline)
mammoth==1.6.0
python-docx>=1.1.0
openpyxl>=3.1.0

# AI / LLM
openai>=1.30.0
anthropic>=0.28.0

# Vector database
pinecone>=4.0.0

# Token counting (used in embedding uploader)
tiktoken>=0.7.0

# Database & auth
flask-sqlalchemy>=3.1.0
PyJWT>=2.8.0

# Environment & config
python-dotenv>=1.0.0

# Utilities (existing pipeline)
pathlib2==2.3.7
typing-extensions>=4.8.0
requests>=2.31.0
PyYAML>=6.0.0
